venv
Added a virtual environment folder (venv) to the project to isolate the dependencies and packages required for using Streamlit. The venv folder contains the Python interpreter, pip, and other scripts that can be used to install and run the project. To activate the virtual environment, run the following command in the terminal:

source venv/bin/activate

To install the required packages, run the following command in the terminal:

pip install -r requirements.txt

To run the Streamlit app, run the following command in the terminal:

streamlit run app.py

The app will open in a new browser tab and display the interactive web application